mcp-tg-jira-seminarapp/pyproject.toml
[project]
name = "mcp-tg-jira-seminarapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "ruff>=0.11.2",
]
